rentbot_parser/requirements.txt
attrs==23.2.0
beautifulsoup4==4.11.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
pycparser==2.22
python-dotenv==0.21.0
requests==2.28.2
selenium==4.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
psutil